+++ __tests__/http.service.test.ts
@@ -12,7 +12,7 @@ beforeAll(async () => {
   });
 });
 describe("HttpService", () => {
-  describe("authedAlqApi", () => {
+  describe.skip("authedAlqApi", () => {
     it("Should return 200", async () => {
       const response = await service.authedAlqApi.get<any>("/perfil");
       expect(response.status).toBe(200);