+++ src/components/deposits/services.ts
@@ -1,10 +1,13 @@
-import { Deposit, Player } from "@prisma/client";
+import { Deposit, Player, Role } from "@prisma/client";
 import CONFIG from "@/config";
 import { ERR } from "@/config/errors";
 import { DepositsDAO } from "@/db/deposits";
 import { CustomError } from "@/helpers/error/CustomError";
 import { CasinoCoinsService } from "@/services/casino-coins.service";
-import { DepositRequest } from "@/types/request/transfers";
+import {
+  DepositRequest,
+  DepositUpdateRequest,
+} from "@/types/request/transfers";
 import { PlainPlayerResponse, RoledPlayer } from "@/types/response/players";
 import { CoinTransferResult, DepositResult } from "@/types/response/transfers";
 import { hidePassword } from "@/utils/auth";
@@ -12,6 +15,7 @@ import { HttpService } from "@/services/http.service";
 import { AlqMovementResponse } from "@/types/response/alquimia";
 import { logtailLogger } from "@/helpers/loggers";
 import { ResourceService } from "@/services/resource.service";
+import { BanxicoService } from "@/services/banxico.service";
 
 export class DepositServices extends ResourceService {
   constructor() {
@@ -34,6 +38,17 @@ export class DepositServices extends ResourceService {
 
     return await this.finalizeDeposit(deposit);
   }
+
+  async update(
+    agent: Player & { roles: Role[] },
+    deposit_id: string,
+    request: DepositUpdateRequest,
+  ): Promise<Deposit> {
+    await DepositsDAO.authorizeUpdate(deposit_id, agent);
+
+    return await DepositsDAO.update(deposit_id, { ...request, dirty: false });
+  }
+
   /**
    * Player announces they have completed a pending deposit
    */
@@ -54,29 +69,36 @@ export class DepositServices extends ResourceService {
   }
 
   /**
-   * Verify a deposit and send coins to player if deposit confirmed
+   * @throws DepositResult
    */
-  private async finalizeDeposit(
+  async verify(
+    deposit: Deposit & { Player: Player },
+  ): Promise<Deposit & { Player: Player }> {
+    if (
+      deposit.status === CONFIG.SD.DEPOSIT_STATUS.VERIFIED ||
+      deposit.status === CONFIG.SD.DEPOSIT_STATUS.CONFIRMED
+    )
+      return deposit;
+
+    const amount = await this.verifyThroughBanxico(deposit);
+    if (!amount) {
+      throw {
+        error: "Deposito no confirmado",
+        deposit: await this.markAsPending(deposit),
+      } as DepositResult;
+    }
+    return await this.markAsVerified(deposit, amount);
+  }
+
+  async transfer(
     deposit: Deposit & { Player: Player },
   ): Promise<DepositResult> {
-    if (
-      deposit.status !== CONFIG.SD.DEPOSIT_STATUS.VERIFIED &&
-      deposit.status !== CONFIG.SD.DEPOSIT_STATUS.CONFIRMED
-    ) {
-      const amount = await this.verify(deposit);
-      if (!amount) {
-        return {
-          error: "Deposito no confirmado",
-          deposit: await this.markAsPending(deposit),
-        };
-      }
-      deposit = await this.markAsVerified(deposit, amount);
-    }
     let coinTransferResult: CoinTransferResult = { ok: false };
     try {
       const casinoCoinsService = new CasinoCoinsService();
       coinTransferResult = await casinoCoinsService.agentToPlayer(deposit);
       if (coinTransferResult.ok) deposit = await this.markAsCompleted(deposit);
+      else deposit = await DepositsDAO.update(deposit.id, { dirty: false });
     } catch (e) {
       if (CONFIG.LOG.LEVEL === "debug") console.error(e);
       if (e instanceof CustomError && e.code === ERR.TRANSACTION_LOG.code)
@@ -94,9 +116,25 @@ export class DepositServices extends ResourceService {
   }
 
   /**
-   * Verify receipt of Player's deposit.
+   * Verify a deposit and send coins to player if deposit confirmed
    */
-  public async verify(deposit: Deposit): Promise<number | undefined> {
+  async finalizeDeposit(
+    deposit: Deposit & { Player: Player },
+  ): Promise<DepositResult> {
+    try {
+      deposit = await this.verify(deposit);
+      return await this.transfer(deposit);
+    } catch (e) {
+      return e as DepositResult;
+    }
+  }
+
+  /**
+   * Verify receipt of Player's deposit through Alquimia.
+   */
+  public async verifyThroughAlquimia(
+    deposit: Deposit,
+  ): Promise<number | undefined> {
     if (
       deposit.status === CONFIG.SD.DEPOSIT_STATUS.VERIFIED ||
       deposit.status === CONFIG.SD.DEPOSIT_STATUS.CONFIRMED
@@ -135,6 +173,41 @@ export class DepositServices extends ResourceService {
     return movements.data.find((movement) => movement.valor_real > 0);
   }
 
+  /**
+   * Verify receipt of Player's deposit through Banxico.
+   * @throws DepositResult
+   */
+  public async verifyThroughBanxico(
+    deposit: Deposit,
+  ): Promise<number | undefined> {
+    if (
+      deposit.status === CONFIG.SD.DEPOSIT_STATUS.VERIFIED ||
+      deposit.status === CONFIG.SD.DEPOSIT_STATUS.CONFIRMED
+    )
+      return deposit.amount!;
+
+    const banxicoService = new BanxicoService();
+
+    if (deposit.sending_bank === "-1") {
+      const bank = await banxicoService.findBank(deposit);
+      if (!bank)
+        throw {
+          error: "Banco no encontrado",
+          deposit: await this.markAsPending(deposit),
+        } as DepositResult;
+      deposit = await DepositsDAO.update(deposit.id, { sending_bank: bank });
+    }
+
+    try {
+      const amount = await banxicoService.verifyDeposit(deposit);
+
+      return amount;
+    } catch (e) {
+      if (CONFIG.LOG.LEVEL === "debug") console.error(e);
+      return;
+    }
+  }
+
   private markAsPending(
     deposit: Deposit,
   ): Promise<Deposit & { Player: Player }> {
@@ -149,7 +222,6 @@ export class DepositServices extends ResourceService {
     amount: number,
   ): Promise<Deposit & { Player: Player }> {
     return DepositsDAO.update(deposit.id, {
-      dirty: false,
       status: CONFIG.SD.DEPOSIT_STATUS.VERIFIED,
       amount,
     });