+++ src/types/response/bankAccount.ts
@@ -2,7 +2,7 @@ export type BankAccount = {
   id: number;
   owner: string;
   player_id: number;
-  bankName: string;
+  bankId: string;
   bankNumber: string;
   bankAlias?: string;
   created_at: string;