+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## [1.0.0] - 2024-07-19
+
+### Changed
+- Replaced `bankName` with `bankId` on BankAccount model.
+
+## [1.0.0] - 2024-07-17
+### Added
+- `/transactions/deposit/:id/update` for agent to mark a deposit as confirmed.
+
+## [1.0.0] - 2024-07-17
+
+### Added
+- Re-added ability to mark payments as paid by agent (meaning payment was made manually).
+
+### Changed
+- Replaced `bankName` with `bankId` on `UserRoot.bankAccount`.
+
+## [1.0.0] - 2024-07-16
+
+### Changed
+- Added `amount`, `date` and `sending_bank` to create deposit request.
+
 ## [1.0.0] - 2024-06-24
 
 ### Changed